\name{chemut-package}
\alias{chemut-package}
\alias{chemut}
\docType{package}
\title{
\packageTitle{chemut}
}
\description{
\packageDescription{chemut}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{chemut}
\packageIndices{chemut}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{chemut}

Maintainer: \packageMaintainer{chemut}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
