% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indel_chr2.R
\name{indel_chr2}
\alias{indel_chr2}
\title{Annotate Indels with masked chr2}
\usage{
indel_chr2(input_file = NULL, ref_file = system.file("extdata", package =
  "chemut"), parental_strain = NULL, mutant_strain = NULL, read_depth = 5,
  var_thld = 0.2, ref_thld = 0.8, freq_diff = 0.4)
}
\arguments{
\item{input_file}{file path of the vcf file (e.g. input_file = "./test.vcf")}

\item{parental_strain}{the name of the parental strain. The name should not contain -. (e.g. parental_strain = "par")}

\item{mutant_strain}{please use a prefix to name mutant strains. Specify the common prefix here (e.g. mutant_strain = "mut" given mutant strains named as mut1, mut2, etc.)}

\item{read_depth}{at a given variant, the read depth (coverage) must be >= read_depth. Default value = 5.}

\item{var_thld}{if the reference allele was detected in <= var_thld (default value = 0.2) of the reads, it is
considered as having a variant allele.}

\item{ref_thld}{if the reference allele was detected in >= ref_thld (default value = 0.8) of the reads, it is 
considered as having a reference allele.}

\item{freq_diff}{at a given variant, the porportion of reads supports the reference allele in the parental
and the mutant strain should differ >= freq_diff (default value = 0.8).}
}
\value{
None
}
\description{
Annotate Indels with masked chr2
}
\examples{
indel(input_file = "/path_to_the_folder/snv.vcf", ref_file = "/path_to_the_folder_of_reference_files/",
     parental_strain = "AX4", mutant_strain = "mutant")

}
\author{
Cheng-Lin Frank Li, \email{chenglinfli@gmail.com}
}

